+++ components/HorizontalTabBar.tsx
@@ -27,6 +27,7 @@ interface Props {
   onClick?: Function;
   containerStyle?: any;
   tabStyle?: any;
+  variant: "underline" | "text";
 }
 
 const HorizontalTabBar = ({
@@ -34,6 +35,7 @@ const HorizontalTabBar = ({
   onClick,
   containerStyle,
   tabStyle,
+  variant,
 }: Props) => {
   const tabContainerEl = useRef<HTMLDivElement>(null);
   const [showRightArrow, setShowRightArrow] = useState(false);
@@ -64,7 +66,14 @@ const HorizontalTabBar = ({
       key: value,
       className: css(
         styles.tab,
-        isSelected ? styles.tabSelected : styles.tabNotSelected,
+        isSelected &&
+          variant === "underline" &&
+          styles.underlineVariantSelected,
+        !isSelected &&
+          variant === "underline" &&
+          styles.underlineVariantNotSelected,
+        isSelected && variant === "text" && styles.textVariantSelected,
+        !isSelected && variant === "text" && styles.textVariantNotSelected,
         tabStyle
       ),
       ...(href && { href }),
@@ -197,17 +206,30 @@ const styles = StyleSheet.create({
     alignItems: "center",
     columnGap: "8px",
   },
-  tabNotSelected: {
+  underlineVariant: {
+    borderBottom: `solid 3px ${colors.GREY()}`,
     ":hover": {
-      color: colors.MEDIUM_GREY(),
       borderBottom: `solid 3px ${colors.GREY()}`,
       transition: "all 0.2s ease-in-out",
     },
   },
-  tabSelected: {
-    color: colors.NEW_BLUE(),
+  underlineVariantSelected: {
     borderBottom: "solid 3px",
     borderColor: colors.NEW_BLUE(),
+  },
+  underlineVariantNotSelected: {
+    ":hover": {
+      color: colors.MEDIUM_GREY(),
+      transition: "all 0.2s ease-in-out",
+    },
+  },
+  textVariantSelected: {
+    color: colors.NEW_BLUE(),
+  },
+  textVariantNotSelected: {
+    ":hover": {
+      color: colors.NEW_BLUE(),
+    },
   },
 });
