+++ components/LiveFeed/LiveFeed.tsx
@@ -1,5 +1,11 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faFlag } from "@fortawesome/pro-solid-svg-icons";
+import {
+  faFlag,
+  faComments,
+  faStar,
+  faCircleCheck,
+  faGrid2,
+} from "@fortawesome/pro-solid-svg-icons";
 import fetchContributionsAPI, { ApiFilters } from "./api/fetchContributionsAPI";
 import {
   CommentContributionItem,
@@ -18,25 +24,177 @@ import ContributionEntry from "./Contribution/ContributionEntry";
 import { flagGrmContent } from "../Flag/api/postGrmFlag";
 import LiveFeedCardPlaceholder from "~/components/Placeholders/LiveFeedCardPlaceholder";
 import Link from "next/link";
+import ResearchCoinIcon from "../Icons/ResearchCoinIcon";
+import HorizontalTabBar, { Tab } from "~/components/HorizontalTabBar";
+import { PaperIcon } from "~/config/themes/icons";
+import { faGlobe, faX } from "@fortawesome/pro-regular-svg-icons";
+
+import BaseModal from "../Modals/BaseModal";
+// import { css, StyleSheet } from "aphrodite";
+// import VerificationForm from "./VerificationForm";
+import { breakpoints } from "~/config/themes/screen";
+import { Hub } from "~/config/types/hub";
+import { useRouter } from "next/router";
+import { getHubs } from "~/components/Hubs/api/fetchHubs";
+import HubSelect from "../Hubs/HubSelect";
+
+const HubSelectModal = ({
+  isModalOpen = true,
+  handleModalClose,
+  handleSelect,
+}) => {
+  const [hubs, setHubs] = useState<Array<Hub>>([]);
+
+  useEffect(() => {
+    (async () => {
+      // @ts-ignore
+      const { hubs, count } = await getHubs({});
+      setHubs(hubs);
+    })();
+  }, []);
+
+  return (
+    <BaseModal
+      offset={"0px"}
+      isOpen={isModalOpen}
+      hideClose={false}
+      title={"Filter by Hub"}
+      closeModal={handleModalClose}
+      zIndex={12}
+      modalStyle={styles1.modalStyle}
+      modalContentStyle={styles1.modalContentStyle}
+    >
+      <div className={css(styles1.formWrapper)}>
+        <HubSelect
+          count={1}
+          hubs={hubs}
+          handleClick={(hub) => {
+            handleSelect(hub);
+          }}
+        />
+      </div>
+    </BaseModal>
+  );
+};
+
+const styles1 = StyleSheet.create({
+  formWrapper: {
+    width: 540,
+    height: "100%",
+    [`@media only screen and (max-width: ${breakpoints.small.str})`]: {
+      width: "100%",
+    },
+    [`@media only screen and (max-width: ${breakpoints.xxsmall.str})`]: {
+      width: "90%",
+    },
+  },
+  modalStyle: {
+    width: "540px",
+    maxHeight: 600,
+  },
+  modalTitleStyleOverride: {},
+  modalContentStyle: {
+    position: "relative",
+    minHeight: 560,
+    padding: "50px 25px ",
+    [`@media only screen and (max-width: ${breakpoints.xxsmall.str})`]: {
+      height: "100%",
+    },
+  },
+  prevActionWrapper: {
+    position: "absolute",
+    top: 12,
+    left: 10,
+  },
+});
+
+export const tabs: Array<Tab> = [
+  {
+    icon: <FontAwesomeIcon icon={faGlobe} />,
+    label: "All",
+    value: "all",
+    isSelected: true,
+  },
+  {
+    icon: <FontAwesomeIcon icon={faComments} />,
+    label: "Conversation",
+    value: "conversation",
+  },
+  {
+    icon: (
+      <ResearchCoinIcon
+        version={4}
+        color={colors.BLACK(0.5)}
+        height={14}
+        width={14}
+      />
+    ),
+    selectedIcon: (
+      <ResearchCoinIcon
+        version={4}
+        color={colors.NEW_BLUE(1.0)}
+        height={14}
+        width={14}
+      />
+    ),
+    label: "Bounties",
+    value: "bounties",
+  },
+  {
+    icon: <FontAwesomeIcon icon={faStar} />,
+    label: "Peer Reviews",
+    value: "reviews",
+  },
+  {
+    icon: <PaperIcon height={14} width={14} onClick={undefined} />,
+    label: "Articles",
+    value: "articles",
+  },
+];
 
 export default function LiveFeed({ hub, isHomePage }): ReactElement<"div"> {
+  const router = useRouter();
   const [appliedFilters, setAppliedFilters] = useState<ApiFilters>({
     hubId: hub?.id as ID,
+    contentType: "all",
   });
   const [isLoadingMore, setIsLoadingMore] = useState<boolean>(false);
   const [isLoadingPage, setIsLoadingPage] = useState<boolean>(true);
 
   const [results, setResults] = useState<Array<Contribution>>([]);
   const [nextResultsUrl, setNextResultsUrl] = useState<any>(null);
+  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
 
   useEffect(() => {
-    let appliedFilters = { hubId: null };
-    if (hub?.id) {
-      appliedFilters = { hubId: hub.id };
-    }
-    setAppliedFilters(appliedFilters);
-    loadResults(appliedFilters, null);
-  }, [hub, isHomePage]);
+    const _appliedFilters: ApiFilters = {
+      // Defaults
+      hubId: null,
+      contentType: "all",
+    };
+    const availableContentTypes = tabs.map((t) => t.value);
+
+    const hasContentTypeFilter =
+      router.query?.contentType &&
+      availableContentTypes.includes(router.query.contentType as string);
+    _appliedFilters.contentType = hasContentTypeFilter
+      ? (router.query.contentType as string)
+      : "all";
+    _appliedFilters.hubId = router.query?.hubId
+      ? (router.query.hubId as string)
+      : null;
+
+    setAppliedFilters(_appliedFilters);
+    loadResults(_appliedFilters, null);
+  }, [router.query]);
+
+  // useEffect(() => {
+  //   let appliedFilters = { hubId: null };
+  //   if (hub?.id) {
+  //     appliedFilters = { hubId: hub.id };
+  //   }
+  //   setAppliedFilters(appliedFilters);
+  //   loadResults(appliedFilters, null);
+  // }, [hub, isHomePage]);
 
   const loadResults = (filters: ApiFilters, url = null) => {
     if (!url) {
@@ -78,81 +236,6 @@ export default function LiveFeed({ hub, isHomePage }): ReactElement<"div"> {
               entry={result}
               actions={[]}
               context="live-feed"
-              // Kobe: Let's expose flagging in the live feed when the UI is ready and supports
-              // actions.
-
-              // actions={[
-              //   {
-              //     html: (
-              //       <FlagButtonV2
-              //         modalHeaderText="Flag Content"
-              //         flagIconOverride={styles.flagIcon}
-              //         iconOverride={
-              //           <FontAwesomeIcon icon={faFlag}></FontAwesomeIcon>
-              //         }
-              //         errorMsgText="Failed to flag"
-              //         successMsgText="Content flagged"
-              //         primaryButtonLabel="Flag"
-              //         subHeaderText="I am flagging this content because of:"
-              //         onSubmit={(
-              //           flagReason,
-              //           renderErrorMsg,
-              //           renderSuccessMsg
-              //         ) => {
-              //           let args: any = {
-              //             flagReason,
-              //             onError: renderErrorMsg,
-              //             onSuccess: renderSuccessMsg,
-              //           };
-
-              //           let item = result.item;
-              //           if (result.contentType.name === "comment") {
-              //             item = item as CommentContributionItem;
-              //             args.commentPayload = {
-              //               ...(result._raw.content_type.name === "thread" && {
-              //                 threadID: item.id,
-              //                 commentType: "thread",
-              //               }),
-              //               ...(result._raw.content_type.name === "comment" && {
-              //                 commentID: item.id,
-              //                 commentType: "comment",
-              //               }),
-              //               ...(result._raw.content_type.name === "reply" && {
-              //                 replyID: item.id,
-              //                 commentType: "reply",
-              //               }),
-              //             };
-              //           }
-
-              //           const unifiedDocument: UnifiedDocument =
-              //             // @ts-ignore
-              //             item.unifiedDocument;
-              //           if (
-              //             ["paper", "post", "hypothesis", "question"].includes(
-              //               unifiedDocument.documentType
-              //             )
-              //           ) {
-              //             args = {
-              //               contentType: unifiedDocument.documentType,
-              //               // @ts-ignore
-              //               contentID: unifiedDocument.document.id,
-              //               ...args,
-              //             };
-              //           } else {
-              //             console.error(
-              //               `${result.contentType.name} Not supported for flagging`
-              //             );
-              //             return false;
-              //           }
-
-              //           flagGrmContent(args);
-              //         }}
-              //       />
-              //     ),
-              //     label: "Flag",
-              //     isActive: true,
-              //   },
-              // ]}
             />
           ),
         };
@@ -178,6 +261,64 @@ export default function LiveFeed({ hub, isHomePage }): ReactElement<"div"> {
 
   return (
     <div className={css(styles.pageWrapper) + " live-feed"}>
+      <div className={css(styles.titleContainer)}>
+        <h1 className={css(styles.title) + " clamp2"}>Live Feed</h1>
+      </div>
+      <div className={css(styles.description)}>
+        Stream of real-time activity on ResearchHub.
+      </div>
+      <HubSelectModal
+        isModalOpen={isModalOpen}
+        handleModalClose={() => setIsModalOpen(false)}
+        handleSelect={(hubs) => {
+          if (hubs?.length > 0) {
+            const hub = hubs[0];
+            setIsModalOpen(false);
+            setAppliedFilters({ hubId: hub.id });
+            router.push(`/live?hubId=${hub.id}`, undefined, { shallow: true });
+          }
+        }}
+      />
+      <div
+        style={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "space-between",
+        }}
+      >
+        <HorizontalTabBar tabs={tabs} variant="text" tabStyle={styles.tab} />
+        <div
+          className={css(
+            styles.tab,
+            styles.hubsFilter,
+            Boolean(appliedFilters.hubId) && styles.hubFilterSelected
+          )}
+          onClick={() => {
+            if (appliedFilters.hubId) {
+              setAppliedFilters({ hubId: null });
+              router.push(`/live`, undefined, { shallow: true });
+            } else {
+              setIsModalOpen(true);
+            }
+          }}
+        >
+          <FontAwesomeIcon icon={faGrid2}></FontAwesomeIcon>
+          Hubs
+          {appliedFilters.hubId && (
+            <div
+              style={{
+                background: "rgb(233 233 233)",
+                borderRadius: "5px",
+                padding: "2px 10px",
+                color: colors.BLACK(1.0),
+                fontSize: 12,
+              }}
+            >
+              <FontAwesomeIcon icon={faX}></FontAwesomeIcon>
+            </div>
+          )}
+        </div>
+      </div>
       {isLoadingPage ? (
         <div className={css(styles.placeholderWrapper)}>
           {Array(10)
@@ -212,6 +353,43 @@ export default function LiveFeed({ hub, isHomePage }): ReactElement<"div"> {
 }
 
 const styles = StyleSheet.create({
+  description: {
+    fontSize: 15,
+    marginBottom: 15,
+    maxWidth: 790,
+    lineHeight: "22px",
+  },
+  title: {
+    fontSize: 30,
+    fontWeight: 500,
+    textOverflow: "ellipsis",
+    marginBottom: 0,
+  },
+  titleContainer: {
+    alignItems: "center",
+    display: "flex",
+    width: "100%",
+    marginBottom: 15,
+  },
+  tab: {
+    fontSize: 14,
+  },
+  hubFilterSelected: {
+    color: colors.NEW_BLUE(1),
+  },
+  hubsFilter: {
+    fontWeight: 500,
+    fontSize: 14,
+    display: "flex",
+    columnGap: "5px",
+    alignItems: "center",
+    borderBottom: "3px solid transparent",
+    padding: "1rem 10px",
+    ":hover": {
+      color: colors.NEW_BLUE(1),
+      cursor: "pointer",
+    },
+  },
   pageWrapper: {
     maxWidth: 800,
     width: 800,
