+++ components/LiveFeed/api/fetchContributionsAPI.ts
@@ -4,7 +4,8 @@ import { captureEvent } from "~/config/utils/events";
 import { ID } from "~/config/types/root_types";
 
 export type ApiFilters = {
-  hubId?: ID;
+  hubId?: ID | null;
+  contentType?: string | null;
 };
 
 type Args = {
